Revert "Merge branch 'master' of https://github.com/silferysky/RogueArcher"
This reverts commit 9544ec2d3758ad97ad8c2e403c3e8943c2052793, reversing
changes made to 58471c922687bffd1eb336d3e0eeb248eae82b9e.

+++ RogueEngine/Source/EntityManager.h
@@ -35,7 +35,7 @@ namespace Rogue
 			// Take an ID from the front of the queue
 			if (REActiveEntityCount > MAX_ENTITIES)
 			{
-				REActiveEntityCount = static_cast<uint32_t>(m_currentActiveObjects.size());
+				REActiveEntityCount = m_currentActiveObjects.size();
 			}
 			RE_ASSERT(REActiveEntityCount < MAX_ENTITIES, "Too many entities in existence.");
 			Entity id = REAvailableEntities.front();
